+++ modules/backend/formwidgets/richeditor/assets/js/richeditor.js
@@ -36,6 +36,7 @@
 
     RichEditor.DEFAULTS = {
         linksHandler: null,
+        uploadHandler: null,
         stylesheet: null,
         fullpage: false,
         editorLang: 'en',
@@ -79,6 +80,7 @@
             language: this.options.editorLang,
             fullPage: this.options.fullpage,
             pageLinksHandler: this.options.linksHandler,
+            uploadHandler: this.options.uploadHandler,
             aceEditorVendorPath: this.options.aceVendorPath,
             toolbarSticky: false
         }
@@ -153,13 +155,27 @@
 
         froalaOptions.shortcutsEnabled = ['show', 'bold', 'italic', 'underline', 'indent', 'outdent', 'undo', 'redo']
 
+        // Ensure that October recognizes AJAX requests from Froala
+        froalaOptions.requestHeaders = {
+            'X-CSRF-TOKEN': $('meta[name="csrf-token"]').attr('content'),
+            'X-Requested-With': 'XMLHttpRequest'
+        }
+
+        // Get the data from the parent form for including in the request
+        var $form = this.$el.closest('form')
+        var formData = {};
+        if ($form.length > 0) {
+            $.each($form.serializeArray(), function (index, field) {
+                formData[field.name] = field.value;
+            })
+        }
+
         // File upload
         froalaOptions.imageUploadURL = froalaOptions.fileUploadURL = window.location
         froalaOptions.imageUploadParam = froalaOptions.fileUploadParam = 'file_data'
-        froalaOptions.imageUploadParams = froalaOptions.fileUploadParams = {
-            X_OCTOBER_MEDIA_MANAGER_QUICK_UPLOAD: 1,
-            _token: $('meta[name="csrf-token"]').attr('content')
-        }
+        froalaOptions.imageUploadParams = froalaOptions.fileUploadParams = $.extend(formData, {
+            _handler: froalaOptions.uploadHandler,
+        })
 
         var placeholder = this.$textarea.attr('placeholder')
         froalaOptions.placeholderText = placeholder ? placeholder : ''
