+++ modules/backend/widgets/mediamanager/partials/_toolbar.htm
@@ -4,8 +4,7 @@
             <div data-control="toolbar">
                 <?php if (!$this->readOnly) : ?>
                     <div class="btn-group offset-right">
-                        <button type="button" class="btn btn-primary oc-icon-upload" data-control="upload"
-                            ><?= e(trans('backend::lang.media.upload')) ?></button>
+                        <button type="button" class="btn btn-primary oc-icon-upload" data-control="upload"><?= e(trans('backend::lang.media.upload')) ?></button>
                         <button type="button" class="btn btn-primary oc-icon-folder" data-command="create-folder"><?= e(trans('backend::lang.media.add_folder')) ?></button>
                     </div>
                 <?php endif; ?>
@@ -14,10 +13,8 @@
 
                 <?php if (!$this->readOnly) : ?>
                     <div class="btn-group offset-right">
-                        <button type="button" class="btn btn-default oc-icon-reply-all" data-command="move"
-                            ><?= e(trans('backend::lang.media.move')) ?></button>
-                        <button type="button" class="btn btn-default oc-icon-trash" data-command="delete"
-                            ><?= e(trans('backend::lang.media.delete')) ?></button>
+                        <button type="button" class="btn btn-default oc-icon-reply-all" data-command="move"><?= e(trans('backend::lang.media.move')) ?></button>
+                        <button type="button" class="btn btn-default oc-icon-trash" data-command="delete"><?= e(trans('backend::lang.media.delete')) ?></button>
                     </div>
                 <?php endif; ?>
